% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeStreams.R
\name{analyzeStreams}
\alias{analyzeStreams}
\title{Title}
\usage{
analyzeStreams(pct_threshold_down = 1.1, pct_threshold_up = 0.9,
  NHDFlowline_folder = NULL, gages_filename = NULL,
  to_from_COMIDs_filename = NULL, valid_artificial_segments_filename = NULL,
  breakpoint_COMIDs = NULL)
}
\arguments{
\item{valid_artificial_segments_filename}{}
}
\description{
Title
}
