% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{Title}
\usage{
loadData(shapefile_folder = NULL, gages_filename = NULL,
  flow_ids_filename = NULL, valid_artificial_segments_filename = NULL)
}
\arguments{
\item{valid_artificial_segments_filename}{}
}
\description{
Title
}
