% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessData.R
\name{preprocessData}
\alias{preprocessData}
\title{Title}
\usage{
preprocessData(segment_data_original, gages, flow_ids, valid_artifical_segments)
}
\arguments{
\item{valid_artifical_segments}{}
}
\description{
Title
}
