% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeData.R
\name{writeData}
\alias{writeData}
\title{Title}
\usage{
writeData(output_data, csv_filename = NULL, ogr_folder = NULL)
}
\arguments{
\item{output_data}{Data to write out}

\item{csv_filename}{Filename to write out the dataframe}

\item{ogr_folder}{Folder to write the OGR shape files}
}
\description{
Title
}
